Fixed problem where if a finger paused the gesture recognition but didn't result in a sound being played, the other fingers would not play. Now if the first finger that triggers a gesture doesn't play a sound, the other fingers are checked for position.
Implemented MIDI like messages to pass to the audio classes. Messages include velocity (0-127), which drum is hit left/right, and how far away from the drum center it is (0-127) with 0 being the center and 127 the edge.
Also changed to using JSyn rather than Beads as found that results were faster after experiments.